ESFIX: Fix zombie queue left behind
1. There is zombie queue left after some service restarted
which has the name likes
st2.sensor.watch.sensor_container-<random_string>,
so set auto_delete as True when initializing Queue instance
of kombu to solve the problem.
The link of the problem in community is:
https://github.com/StackStorm/st2/issues/3811
2. Add pep8 in dozer.
3. There are some errors in the output of log:
Double requirement given: nose>=1.3.7
Double requirement given: mock==2.0.0
the reason is that nose and mock both are
in requirements.txt and test-requirements.txt,
so it needs to remove the repeated mock and nose
in test-requirements.txt.

Change-Id: I1d869d90fe092c274ef36b6cffbc1b565d17b712
Closes-Task:#EAS-23301