Fix retry to connect rabbitmq failed
The st2rulesengine service can not connect to
rabbitmq if rabbitmq stoped and restarted after
some time, it will raise error:
failed to resolve broker hostname
error: [Errno 104] ECONNRESET
There are two reasone about it:
Firstly,  the max_retries is not set which will
result in infinnite loop when it rabbitmq failed.
Secondly, the eventlet and kombu in v2.6 version
of dozer used need to update version because
there are some bugs in these old version of kombu
and eventlet which also makes this problem occred.
So it needs to set max retry times and update
version of eventlet and kombu to solve this problem.

Closes-Bug:#EAS-32121

Change-Id: Ie21941137df5cbc52630b09d9d3928fd101de855